% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPstrat_poisson_NH.R
\name{bic_mse_lambda_simp}
\alias{bic_mse_lambda_simp}
\title{bic_mse_lambda_simp}
\usage{
bic_mse_lambda_simp(df, y, vars, lambda_grid)
}
\arguments{
\item{df}{data frame with variables of interest}

\item{y}{response variable (poisson distributed)}

\item{vars}{character vector of variable names to include in the model}

\item{lambda_grid}{grid of lambda values to test}
}
\value{
a list containing the optimal lambda value, BIC values, and MSE values
}
\description{
Find the optimal lambda value over a grid using BIC
}
\examples{
data("data")
bic_mse_lambda_simp(data, 'hospital_days', vars = c('age_group', 'pulm_impair'), seq(0.1, 1, 0.1))
}
