% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPstrat_poisson_NH.R
\name{bic_poisson}
\alias{bic_poisson}
\title{bic_poisson}
\usage{
bic_poisson(df, y, var1, var2, beta)
}
\arguments{
\item{df}{data frame with variables of interest}

\item{y}{response variable (poisson distributed)}

\item{var1}{first categorical variable}

\item{var2}{second categorical variable}

\item{beta}{vector of model coefficients}
}
\value{
BIC value
}
\description{
Find the BIC for a given LASSO tree interaction model
}
\examples{
data("data")
bic_poisson(data, 'hospital_days', var1 = 'age_group', var2 = 'pulm_impair', beta = rep(1, 28))
}
