% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPstrat_poisson_NH.R
\name{bic_poisson_simp}
\alias{bic_poisson_simp}
\title{bic_poisson_simp}
\usage{
bic_poisson_simp(df, y, vars, beta)
}
\arguments{
\item{df}{data frame with variables of interest}

\item{y}{response variable (poisson distributed)}

\item{vars}{character vector of variable names to include in the model}

\item{beta}{vector of model coefficients}
}
\value{
BIC value
}
\description{
Find the BIC for a given LASSO tree interaction model
}
\examples{
data("data")
bic_poisson_simp(data, 'hospital_days', c('age_group', 'pulm_impair'), beta = rep(1, 10))
}
