% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPstrat_poisson_NH.R
\name{get_A}
\alias{get_A}
\title{get_A}
\usage{
get_A(nlevels_vec)
}
\arguments{
\item{nlevels_vec}{This is an integer vector containing the number of levels for each categorical variable}
}
\value{
An n x p matrix of constraints, where n is the number of constraints, and p is the length of beta
}
\description{
Creates the matrix of constraints to be passed into quadratic programming problem for non-interaction model
}
\examples{
get_A(c(7, 4))
}
