% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPstrat_poisson_NH.R
\name{get_Ax}
\alias{get_Ax}
\title{get_Ax}
\usage{
get_Ax(p, q)
}
\arguments{
\item{p}{number of levels for the first categorical variable}

\item{q}{number of levels for the second categorical variable}
}
\value{
An n x p matrix of constraints, where n is the number of constraints, and p is the length of beta
}
\description{
Creates the matrix of constraints to be passed into quadratic programming problem for interaction model
}
\examples{
get_Ax(7, 4)
}
