% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPstrat_poisson_NH.R
\name{penalty_vec}
\alias{penalty_vec}
\title{penalty_vec}
\usage{
penalty_vec(p, q)
}
\arguments{
\item{p}{number of levels for the first categorical variable}

\item{q}{number of levels for the second categorical variable}
}
\value{
A vector of length beta that when multiplied by beta gives the LASSO penalty
}
\description{
Creates the matrix of constraints to be passed into quadratic programming problem for the interaction model
}
\examples{
penalty_vec(7, 4)
}
