% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPstrat_poisson_NH.R
\name{penalty_vec_simp}
\alias{penalty_vec_simp}
\title{penalty_vec_simp}
\usage{
penalty_vec_simp(nlevels_vec)
}
\arguments{
\item{nlevels_vec}{This is an integer vector containing the number of levels for each categorical variable}
}
\value{
a vector of length beta that when multiplied by beta gives the LASSO penalty
}
\description{
Creates the penalty vector for the non-interaction model
}
\examples{
penalty_vec_simp(c(7,4))
}
