% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QPstrat_poisson_NH.R
\name{strat_poissonX}
\alias{strat_poissonX}
\title{strat_poissonX}
\usage{
strat_poissonX(
  df,
  y,
  var1,
  var2,
  lambda = 0,
  tol = 1e-06,
  max_iter = 100,
  warm.start
)
}
\arguments{
\item{df}{data frame with variables of interest}

\item{y}{response variable (poisson distributed)}

\item{var1}{first categorical variable}

\item{var2}{second categorical variable}

\item{lambda}{tuning parameter for the LASSO penalty}

\item{tol}{tolerance level for convergence}

\item{max_iter}{maximum number of iterations}

\item{warm.start}{initial values for beta}
}
\value{
a vector containing the beta coefficients for the interaction grid
}
\description{
Fits the LASSO tree model on the given data assuming an interaction grid
}
\examples{
data("data")
strat_poissonX(data, 'hospital_days', var1 = 'age_group', var2 = 'pulm_impair', lambda = 0)
}
